feat: Ensure and compose connection process
This commit composes the server connection process by
defining a series of steps until the connection to the server
is either accepted or rejected. When rejected the right logging is
done and when accepted debud messages are printed to the console.

The steps look as follows:

spawn (asls -p) -> connect -> resolve